+++ src/components/Footer.tsx
@@ -1,8 +1,12 @@
 import { Instagram, Facebook, Twitter } from "lucide-react";
 
-export const Footer = () => {
+interface FooterProps {
+  className?: string;
+}
+
+export const Footer: React.FC<FooterProps> = ({ className = "" }) => {
   return (
-    <footer className="mt-36 py-12 mb-3">
+    <footer className={`py-12 mb-3 ${className}`}>
       <div className="container mx-auto flex flex-col md:flex-row justify-between items-center gap-8 px-6">
         <div className="flex flex-col items-center md:items-start">
           <p className="text-3xl font-extrabold text-gray-800">Хурдан хоол</p>
